Isolate createTransfer onSubmit part that creates state to make it testable
