+++ packages/browser-wallet/src/popup/pages/Account/SendCcd/CreateTransfer.tsx
@@ -6,26 +6,14 @@ import { selectedAccountAtom } from '@popup/store/account';
 import Form from '@popup/shared/Form';
 import AmountInput from '@popup/shared/Form/AmountInput';
 import Input from '@popup/shared/Form/Input';
-import {
-    ccdToMicroCcd,
-    getPublicAccountAmounts,
-    fractionalToInteger,
-    useAsyncMemo,
-    integerToFractional,
-    max,
-    displayAsCcd,
-} from 'wallet-common-helpers';
+import { getPublicAccountAmounts, useAsyncMemo, integerToFractional, max, displayAsCcd } from 'wallet-common-helpers';
 import { AccountAddress, SimpleTransferPayload } from '@concordium/web-sdk';
 import { SubmitHandler, useForm, Validate } from 'react-hook-form';
 import clsx from 'clsx';
 import { useLocation, useNavigate } from 'react-router-dom';
 
 import Submit from '@popup/shared/Form/Submit';
-import {
-    buildSimpleTransferPayload,
-    validateTransferAmount,
-    validateAccountAddress,
-} from '@popup/shared/utils/transaction-helpers';
+import { validateTransferAmount, validateAccountAddress } from '@popup/shared/utils/transaction-helpers';
 import { useAccountInfo } from '@popup/shared/AccountInfoListenerContext';
 import { useSelectedCredential } from '@popup/shared/utils/account-helpers';
 import { CCD_METADATA } from '@shared/constants/token-metadata';
@@ -45,14 +33,7 @@ import { SIMPLE_TRANSFER_ENERGY_TOTAL_COST } from '@shared/utils/energy-helpers'
 import Img from '@popup/shared/Img';
 import { routes } from './routes';
 import PickToken from './PickToken';
-
-export type FormValues = {
-    amount: string;
-    recipient: string;
-    executionEnergy: string;
-    cost: string;
-    token?: TokenIdentifier;
-};
+import { buildConfirmState, CreateTransferFormValues } from './util';
 
 interface Props {
     exchangeRate?: number;
@@ -60,6 +41,13 @@ interface Props {
     setCost: (cost: bigint) => void;
     setDetailsExpanded: (expanded: boolean) => void;
 }
+
+const getNextRoute = (token?: TokenIdentifier) => {
+    if (token) {
+        return routes.confirmToken;
+    }
+    return routes.confirm;
+};
 
 /**
  * undefined is used to denote "not resolved yet", to align with useAsyncMemo return type
@@ -93,7 +81,7 @@ function CreateTransaction({ exchangeRate, tokens, setCost, setDetailsExpanded, 
     const selectedCred = useSelectedCredential();
     const nav = useNavigate();
     const accountTokens = useMemo(() => (address ? tokens[address] || {} : undefined), [tokens, address]);
-    const form = useForm<FormValues>({
+    const form = useForm<CreateTransferFormValues>({
         defaultValues: createDefaultValues(state as State, accountTokens),
     });
     const client = useAtomValue(jsonRpcClientAtom);
@@ -213,17 +201,10 @@ function CreateTransaction({ exchangeRate, tokens, setCost, setDetailsExpanded, 
         return null;
     }
 
-    const onSubmit: SubmitHandler<FormValues> = (vs) => {
-        if (vs.token) {
-            const payload = buildSimpleTransferPayload(
-                vs.recipient,
-                fractionalToInteger(vs.amount, getMetadataDecimals(vs.token.metadata))
-            );
-            nav(routes.confirmToken, { state: { ...payload, ...vs.token, executionEnergy: vs.executionEnergy } });
-        } else {
-            const payload = buildSimpleTransferPayload(vs.recipient, ccdToMicroCcd(vs.amount));
-            nav(routes.confirm, { state: { ...payload } });
-        }
+    const onSubmit: SubmitHandler<CreateTransferFormValues> = (vs) => {
+        const nextRoute = getNextRoute(vs.token);
+        const currentState = buildConfirmState(vs);
+        nav(nextRoute, { state: currentState });
     };
 
     if (pickingToken) {
